Cleaned up unused code, starting documentation
