+++ .env.example
@@ -1,0 +1,4 @@
+ELASTIC_ENDPOINT_URL=
+FAUNADB_KEY=
+MAILER_USER=
+MAILER_PASS=
