+++ CONTRIBUTING.md
@@ -1,0 +1,12 @@
+## Repo Overview
+
+| Folder | Purpose |
+| ----------- | ----------- |
+| /cache | Houses the local data cache |
+| /email | Outgoing email templates for various purposes |
+| /keys | Middleware for API Key Authentication feature |
+| /public | Assets to be accessible via Public URL |
+| /rest | Contains the very heart and soul of the REST API |
+| /test | Self-explainatory |
+| /views | Text |
+| /utils | Utilities for server maintenance |
