+++ README.md
@@ -1,0 +1,14 @@
+<h1>
+  <img src="./public/images/peterportal-banner-logo.png" alt="PeterPortal" width="400" />
+</h1>
+
+PeterPortal Public API provides software developers with easy-access to UC Irvine publicly availale data such as: courses information, professor information, grade distribution, and schedule of classes.
+
+## Built with
+- ExpressJS
+- GraphQL
+
+## Works with
+- FaunaDB
+- ElasticSearch
+
