+++ package.json
@@ -1,5 +1,5 @@
 {
-  "name": "uci-catalogue-search-api",
+  "name": "peterportal-public-api",
   "version": "0.0.0",
   "private": true,
   "scripts": {
@@ -7,38 +7,27 @@
     "test": "jest"
   },
   "dependencies": {
-    "aws-sdk": "^2.770.0",
-    "cheerio": "^1.0.0-rc.3",
-    "cookie-parser": "~1.4.4",
     "cors": "^2.8.5",
-    "data-api-client": "^1.1.0",
     "debug": "~2.6.9",
     "dotenv": "^8.2.0",
     "ejs": "~2.6.1",
     "email-templates": "^7.1.0",
     "express": "~4.16.1",
     "express-graphql": "^0.11.0",
-    "express-mysql-session": "^2.1.4",
-    "express-session": "^1.17.0",
     "faunadb": "^3.0.1",
     "flat-cache": "^2.0.1",
+    "graphql": "^15.3.0",
     "http-errors": "~1.6.3",
     "morgan": "~1.9.1",
-    "mysql": "^2.18.1",
     "node-fetch": "^2.6.0",
     "nodemailer": "^6.4.11",
     "nodemon": "^2.0.2",
-    "passport": "^0.4.1",
-    "passport-facebook": "^3.0.0",
-    "passport-github": "^1.1.0",
-    "passport-google-oauth": "^2.0.0",
     "swagger-jsdoc": "^4.0.0",
-    "swagger-ui-express": "^4.1.4",
+    "underscore": "^1.11.0",
     "websoc-api": "^0.0.5"
   },
   "devDependencies": {
     "jest": "^26.4.2",
-    "serverless-domain-manager": "^4.2.0",
     "supertest": "^5.0.0"
   }
 }
