--- views/admin_home.ejs
@@ -1,29 +1,0 @@
-<!DOCTYPE html>
-<html>
-  <head>
-    <meta charset="utf-8" />
-    <link rel="icon" href="%PUBLIC_URL%/favicon.png" />
-    <link href="https://fonts.googleapis.com/css?family=Open+Sans&display=swap" rel="stylesheet">     
-    <meta name="viewport" content="width=device-width, initial-scale=1" />
-    <link rel="stylesheet" href="//cdn.jsdelivr.net/npm/semantic-ui@2.4.2/dist/semantic.min.css" />
-    
-    <script src="https://code.jquery.com/jquery-3.5.1.js"></script>
-    <script src="https://cdn.datatables.net/1.10.21/js/jquery.dataTables.min.js"></script>
-    <script src="https://cdn.datatables.net/1.10.21/js/dataTables.semanticui.min.js"></script>
-    <script src="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.3.1/semantic.min.js"></script>
-    <meta
-      name="description"
-      content="The One-Stop-Shop for course discovery at UC Irvine."
-    />
-    <meta name="author" content="ICSSC Project Committee"/>
-    <meta name="keywords" content="UC Irvine, UCI, WebSoC, Rate My Professor, AntAlmanac, Catalogue, course planning, schedule, Antplanner"/>
-
-    <title>PeterPortal</title>
-    
-  </head>
-  <body>
-    <img src="/peterportal-banner-logo.svg">
-    <a href="/users/auth/github">Login Via GitHub</a>
-
-  </body>
-</html>
